+++ src/main/java/step3/domain/LottoNumSetGenerator.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -6,9 +6,9 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import static step3.LottoNum.MAX_LOTTO_NUM;
-import static step3.LottoNum.MIN_LOTTO_NUM;
-import static step3.LottoNumSet.LOTTO_NUM_COUNT;
+import static step3.domain.LottoNum.MAX_LOTTO_NUM;
+import static step3.domain.LottoNum.MIN_LOTTO_NUM;
+import static step3.domain.LottoNumSet.LOTTO_NUM_COUNT;
 
 public class LottoNumSetGenerator {
     private static final List<LottoNum> ALL_LOTTO_NUM =
